+++ heatload/main.cpp
@@ -6,19 +6,9 @@
 #include <thread>
 
 #include "adios2.h"
-#include "flags.hpp"
-#include "load.hpp"
-#include "particles.hpp"
-#include "sml.hpp"
-#include "particles.hpp"
 #include "heatload.hpp"
-#include "output.hpp"
 
 #define GET(X, i, j) X[i * 9 + j]
-
-void heatload();
-void init(); // initialization
-void heatload_calc(const Particles &div, HeatLoad &sp, t_ParticleDB &db); // calculate heatload
 
 // extern "C" void set_test_type(int test_type);
 
@@ -49,80 +39,5 @@ int main(int argc, char *argv[]) {
     ad = &adios;
     // run actual routine
     heatload();
-}
-
-void heatload() {
-
-    // init simulation parameters
-    init();
-
-    // init adios
-    load_init("xgc.escaped_ptls.bp");
-
-    t_ParticleDB iesc_db; 
-    t_ParticleDB eesc_db; 
-
-    int i = 0;
-    while (1) {
-        i++;
-
-        Particles idiv;
-        Particles ediv;
-        t_ParticlesList iesc;
-        t_ParticlesList eesc;
-
-        adios2::StepStatus status = load_data(idiv, ediv, iesc, eesc);
-        if (status == adios2::StepStatus::EndOfStream)
-        { 
-            std::cout << "Input stream terminated. Exit loop" << std::endl;
-            break;
-        }
-        else if (status == adios2::StepStatus::NotReady)
-        {
-            std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-            continue;
-        }
-        else if (status == adios2::StepStatus::OtherError)
-        {
-            std::cout << "Input stream had errors. Exit loop" << std::endl;
-            break;
-        }
-
-        std::cout << std::endl;
-        std::cout << ">>> Step: " << i << std::endl;
-        std::cout << "Num. of escaped ions: " << iesc.size() << std::endl;
-        std::cout << "Num. of escaped elec: " << eesc.size() << std::endl;
-        std::cout << "Num. of divertor ions: " << idiv.size() << std::endl;
-        std::cout << "Num. of divertor elec: " << ediv.size() << std::endl;
-
-        // print first 10 esc particles
-        int count = 0;
-        std::map<long long, Particle>::iterator it;
-        for (it = iesc.begin(); it != iesc.end(); it++) {
-            printf("iesc gid, rzphi, flag: %lld %f %f %f %d\n", it->second.gid, it->second.r, it->second.z, it->second.phi, it->second.flag);
-            count++;
-            if (count>10) break;
-        }
-
-
-        // separate divertor particles and escaped particles
-        iesc_db.push_back(iesc);
-        eesc_db.push_back(eesc);
-        Particle ptl = search(iesc_db, i-1, 15824414);
-        printf ("Found or not? gid=%lld\n", ptl.gid);
-
-        // store escaped particles to DB
-
-        // Calculate heatload from divertor particles
-        HeatLoad ion(1);
-        HeatLoad elec(0);
-
-        heatload_calc(idiv, ion,  iesc_db); // need to send DB
-        heatload_calc(ediv, elec, eesc_db);
-        output(ion, elec);
-    }
-
-    load_finalize();
-    output_finalize();
 }
