+++ heatload/output.hpp
@@ -1,4 +1,4 @@
-#include "heatload.hpp"
+#include "heatload_calc.hpp"
 
 void output(HeatLoad &ion, HeatLoad &elec); // output graphs or data for graphs
 void output_finalize();
